requirements/base.txt
Django>=1.7.1,<1.8
Pillow>=2.8.1
django-grappelli>=2.6.0,<=2.6.9
django-filebrowser==3.5.7
django-adminactions>=0.2
django-easy-maps>=0.9,<=0.9.99
django-extensions>=1.5.5
django-localflavor>=1.0,<2.0
django-mailgun>=0.2.0
feedparser>=5.2.0
geopy>=1.10.0
lxml>=3.0.2
pytz>=2015.4
requests>=2.5.0
beautifulsoup4>=4.3.0
six>=1.9.0
xlrd>=0.9.3
xlwt>=0.7.5
